DCMAW-9264: Move requestBuilder to test utils
